Update: TBB 2.0 20080319.

+++ examples/GettingStarted/sub_string_finder/sub_string_finder_extended.cpp
@@ -84,7 +84,7 @@ class SubStringFinder {
   str(s), max_array(m), pos_array(p) { }										
 };				
 
-int main(size_t argc, char *argv[]) {
+int main(int argc, char *argv[]) {
  task_scheduler_init init;
 
  string str[N] = { string("a"), string("b") };
